+++ frontend/controllers/CompanyController.php
@@ -8,6 +8,7 @@ use app\models\CompanySearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\web\ForbiddenHttpException;
 
 /**
  * CompanyController implements the CRUD actions for Company model.
@@ -74,15 +75,22 @@ class CompanyController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Company();
+        if(Yii::$app->user->can('create_company')){
+            $model = new Company();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->Company_id]);
-        } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->Company_id]);
+            } else {
+                return $this->render('create', [
+                    'model' => $model,
+                ]);
+            }    
+        }else{
+            //throw new ForbiddenHttpException;
+            Yii::$app->session->setFlash('danger', 'You are not authorized!');
+            return $this->redirect(['index']);
         }
+        
     }
 
     /**
@@ -93,14 +101,20 @@ class CompanyController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(Yii::$app->user->can('update_company')){
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->Company_id]);
-        } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->Company_id]);
+            } else {
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
+            }
+        }else{
+            //throw new ForbiddenHttpException;
+            Yii::$app->session->setFlash('danger', 'You are not authorized!');
+            return $this->redirect(['index']);
         }
     }
 
@@ -112,9 +126,15 @@ class CompanyController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(Yii::$app->user->can('delete_company')){
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }else{
+            //throw new ForbiddenHttpException;
+            Yii::$app->session->setFlash('danger', 'You are not authorized!');
+            return $this->redirect(['index']);
+        }
     }
 
     /**
